Ajouter le code pour gérer plusieurs pédictions via le réentrainement ou pas du modèle et générer les résultats correspondants à l'exécution.

+++ streamlit_app/pages/2_📥_Dépot_et_Validation_des_Données_2.py
@@ -2,15 +2,10 @@ import os
 import streamlit as st
 import numpy as np
 import pandas as pd
+import shutil
 from menu import display_menu
 from dependency_manager import check_dependencies
 from utilitaires.Preprocessing_des_donnees import preprocesser_les_donnees_1, preprocesser_les_donnees_2
-
-##from codecarbon import EmissionsTracker
-# Initialiser le suivi des émisions
-## tracker = EmissionsTracker(project_name="Mon_Prog_IA", output_file = "emissions.csv")
-## tracker.start()
-## tracker.end()
 
 # Afficher le menu
 display_menu()
@@ -19,6 +14,7 @@ display_menu()
 path_donnees_raw = "streamlit_app/static/donnees/donnees_raw/"    # Jeu de données par défaut (jeu de données fournis lors de la phase 1 de l'hackathon)
 preprocessing_dir = "streamlit_app/static/donnees/donnees_preprocessees/"
 donnees_a_la_volee_dir = os.path.join(preprocessing_dir, "donnees_a_la_volee/")
+resultats_a_la_volee_dossier = "streamlit_app/resultats/donnees_a_la_volee/"
 
 def show():
     
@@ -27,7 +23,7 @@ def show():
     check_dependencies("Dépot et Validation des Données")
 
     if st.button("Nettoyer les dossiers de validation"):
-        # Check if the directory exists
+        # Vérifier si le dossier existe
         if os.path.exists(donnees_a_la_volee_dir) and os.path.isdir(donnees_a_la_volee_dir):
             # Loop through all files in the directory and remove them
             for file_name in os.listdir(donnees_a_la_volee_dir):
@@ -36,12 +32,29 @@ def show():
                     if os.path.isfile(file_path):
                         os.remove(file_path)
                     elif os.path.isdir(file_path):
-                        shutil.rmtree(file_path)  # If there are subdirectories, remove them
+                        shutil.rmtree(file_path)  # S'il y a un sous dossiers, le supprimer.
                 except Exception as e:
                     print(f"Error deleting {file_path}: {e}")
-            print("All files in 'donnees_a_la_volee/' have been deleted.")
+            print("Tous les fichiers de 'donnees_a_la_volee/' ont été supprimés.")
         else:
-            print("The directory 'donnees_a_la_volee/' does not exist.")
+            print("Le dossier 'donnees_preprocessees/donnees_a_la_volee/' n'existe pas.")
+
+    if st.button("Nettoyer le dossier des resultats"):
+        # Vérifier si le dossier existe
+        if os.path.exists(resultats_a_la_volee_dossier) and os.path.isdir(resultats_a_la_volee_dossier):
+            # Loop through all files in the directory and remove them
+            for file_name in os.listdir(resultats_a_la_volee_dossier):
+                file_path = os.path.join(resultats_a_la_volee_dossier, file_name)
+                try:
+                    if os.path.isfile(file_path):
+                        os.remove(file_path)
+                    elif os.path.isdir(file_path):
+                        shutil.rmtree(file_path)  # S'il y a un sous dossiers, le supprimer.
+                except Exception as e:
+                    print(f"Error deleting {file_path}: {e}")
+            print("Tous les fichiers de 'donnees_a_la_volee/' ont été supprimés.")
+        else:
+            print("Le dossier 'resultats/donnees_a_la_volee/' n'existe pas.")
 
     st.write("Veuillez entrer les données pour les 60 dernières secondes:")
 
@@ -114,9 +127,14 @@ def show():
                 # Store properly formatted data in session state
                 st.session_state.prediction_data = prediction_data_df
 
-                st.success("Les données sont valides.")
                 st.session_state.prediction_effectuee = False
+                dernieres_donnees_deposees = prediction_data_df
+                if ('precedentes_donnees_deposees' not in st.session_state) or ('precedentes_donnees_deposees' in st.session_state and not dernieres_donnees_deposees.equals(st.session_state.precedentes_donnees_deposees)):
+                    st.session_state.nouveau_depot_donnees = True
+                st.session_state.precedentes_donnees_deposees = dernieres_donnees_deposees              
                 st.session_state.valid_depot_donnees = True
+                st.success("Les données sont valides.")
+
         else:
             st.error("Erreur: Aucun fichier n'a été téléchargé.")
